# dockstore-webservice/src/main/resources/openapi3/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: theglobalalliance@genomicsandhealth.org
    name: Dockstore@ga4gh
    url: https://discuss.dockstore.org/t/opening-helpdesk-tickets/1506
  description: "This describes the dockstore API, a webservice that manages pairs\
    \ of Docker images and associated metadata such as CWL documents and Dockerfiles\
    \ used to build those images. Explore swagger.json for a Swagger 2.0 description\
    \ of our API and explore openapi.yaml for OpenAPI 3.0 descriptions."
  license:
    name: Apache License Version 2.0
    url: https://github.com/dockstore/dockstore/blob/develop/LICENSE
  termsOfService: TBD
  title: Dockstore API
  version: 1.11.0-alpha.2-SNAPSHOT
servers:
- description: Current server when hosted on AWS
  url: /api
  variables: {}
- description: When working locally
  url: /
  variables: {}
- description: Production server
  url: https://dockstore.org/api
  variables: {}
- description: Staging server
  url: https://staging.dockstore.org/api
  variables: {}
- description: Nightly build server
  url: https://dev.dockstore.net/api
  variables: {}
...
paths:
...
  /api/ga4gh/v2/extended/tools/entry/_search:
    post:
      description: This endpoint searches the indices for all published tools and
        workflows. Used by utilities that expect to talk to an elastic search endpoint.
      operationId: toolsIndexSearch
      requestBody:
        content:
          '*/*':
            schema:
              type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: An elastic search result.
      summary: Search the tools and workflows indices.
      tags:
      - extendedGA4GH
...
  /metadata/runner_dependencies:
    get:
      description: "Returns the file containing runner dependencies, NO authentication"
      operationId: getRunnerDependencies
      parameters:
      - description: The Dockstore client version
        in: query
        name: client_version
        schema:
          type: string
      - description: "Python version, only relevant for the cwltool runner"
        in: query
        name: python_version
        schema:
          type: string
          default: "3"
      - description: The tool runner
        in: query
        name: runner
        schema:
          type: string
          default: cwltool
          enum:
          - cwltool
      - description: Response type
        in: query
        name: output
        schema:
          type: string
          default: text
          enum:
          - json
          - text
      responses:
        default:
          content:
            application/json:
              schema:
                type: string
          description: The requirements.txt file
      summary: Returns the file containing runner dependencies
      tags:
      - metadata
...
components:
...
  securitySchemes:
    bearer:
      scheme: bearer
      type: http